FIR/UAST: commonize KotlinClassViaConstructorUSimpleReferenceExpression
