+++ uast/uast-kotlin-base/src/org/jetbrains/uast/kotlin/BaseKotlinUastResolveProviderService.kt
@@ -40,6 +40,8 @@ interface BaseKotlinUastResolveProviderService {
 
     fun resolveCall(ktElement: KtElement): PsiMethod?
 
+    fun resolveToClassIfConstructorCall(ktCallElement: KtCallElement, source: UElement): PsiElement?
+
     fun resolveToDeclaration(ktExpression: KtExpression): PsiElement?
 
     fun resolveToType(ktTypeReference: KtTypeReference, source: UElement): PsiType?
