+++ uast/uast-kotlin/src/org/jetbrains/uast/kotlin/expressions/KotlinUSimpleReferenceExpression.kt
@@ -19,21 +19,15 @@ package org.jetbrains.uast.kotlin
 import com.intellij.openapi.util.Key
 import com.intellij.psi.*
 import org.jetbrains.kotlin.descriptors.CallableDescriptor
-import org.jetbrains.kotlin.descriptors.ConstructorDescriptor
 import org.jetbrains.kotlin.descriptors.DeclarationDescriptor
-import org.jetbrains.kotlin.kdoc.psi.impl.KDocName
 import org.jetbrains.kotlin.lexer.KtTokens
 import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.resolve.calls.callUtil.getResolvedCall
 import org.jetbrains.kotlin.resolve.calls.model.ResolvedCall
-import org.jetbrains.kotlin.resolve.calls.tower.NewResolvedCallImpl
-import org.jetbrains.kotlin.resolve.sam.SamConstructorDescriptor
 import org.jetbrains.kotlin.synthetic.SyntheticJavaPropertyDescriptor
 import org.jetbrains.uast.*
 import org.jetbrains.uast.internal.acceptList
-import org.jetbrains.uast.internal.log
 import org.jetbrains.uast.kotlin.internal.DelegatedMultiResolve
-import org.jetbrains.uast.kotlin.internal.multiResolveResults
 import org.jetbrains.uast.visitor.UastVisitor
 
 var PsiElement.destructuringDeclarationInitializer: Boolean? by UserDataProperty(Key.create("kotlin.uast.destructuringDeclarationInitializer"))
@@ -155,44 +149,4 @@ class KotlinUSimpleReferenceExpression(
         override fun resolve(): PsiMethod? = resolveToPsiMethod(sourcePsi, accessorDescriptor)
     }
 
-}
-
-class KotlinClassViaConstructorUSimpleReferenceExpression(
-    override val sourcePsi: KtCallElement,
-    override val identifier: String,
-    givenParent: UElement?
-) : KotlinAbstractUExpression(givenParent), USimpleNameReferenceExpression, KotlinUElementWithType {
-    override val resolvedName: String?
-        get() = (resolved as? PsiNamedElement)?.name
-
-    private val resolved by lazy {
-        when (val resultingDescriptor = sourcePsi.getResolvedCall(sourcePsi.analyze())?.descriptorForResolveViaConstructor()) {
-            is ConstructorDescriptor -> {
-                sourcePsi.calleeExpression?.let { resolveToDeclarationImpl(it, resultingDescriptor.constructedClass) }
-            }
-            is SamConstructorDescriptor ->
-                (resultingDescriptor.returnType?.getFunctionalInterfaceType(this, sourcePsi) as? PsiClassType)?.resolve()
-            else -> null
-        }
-    }
-
-    override fun accept(visitor: UastVisitor) {
-        super<KotlinAbstractUExpression>.accept(visitor)
-    }
-
-    override fun resolve(): PsiElement? = resolved
-
-    override fun asLogString(): String {
-        val resolveStr = when (val resolved = resolve()) {
-            is PsiClass -> "PsiClass: ${resolved.name}"
-            is PsiMethod -> "PsiMethod: ${resolved.name}"
-            else -> resolved.toString()
-        }
-        return log<USimpleNameReferenceExpression>("identifier = $identifier, resolvesTo = $resolveStr")
-    }
-
-    // In new inference, SAM constructor is substituted with a function descriptor, so we use candidate descriptor to preserve behavior
-    private fun ResolvedCall<*>.descriptorForResolveViaConstructor(): CallableDescriptor? {
-        return if (this is NewResolvedCallImpl) candidateDescriptor else resultingDescriptor
-    }
 }
